+++ CHANGELOG.md
@@ -7,8 +7,8 @@
   * 增强 `DateFormatFieldConvertor`，使支持 `Temporal` 及其子类的对象的格式化
   * 增强 `DateFormatFieldConvertor`，新增：`setZoneId(ZoneId)` 方法，可配置时区
   * 新增 `BFieldConvertor` 与 `MFieldConvertor` 字段转换器子接口，提高字段转换效能
-  * 新增 `DefaultDbMapping`，从 `DefaultMetaResolver` 的内部类独立出来
-  * 增强 `DbMapping`，使其完全接管数据映射的解析工作（TODO）
+  * 增强 `DbMapping`，使其完全接管数据映射的解析工作，并将 `DefaultDbMapping` 从 `DefaultMetaResolver` 的内部独立出来
+  * 增强 `DefaultDbMapping`，使支持配置注解缺省时的表名前缀与是否开启大写映射，即支持默认映射大写的表名与列名
 * Bean Searcher Boot Starter
   * 支持配置 `NamedDataSource` 类型的 Bean 来添加多个具名数据源
   * 自动添加 `spring-boot-starter-jdbc` 依赖